Extracted typing indicator tests.
